Agent, orchestrator, memory, and API compatibility updates
Agent compatibility and execution contract updates across base_agent.py, agent_factory.py, and specialized_agents.py.

Financial and translator feature completeness in financial_agent.py, translator_agent.py, and translator.py.

Orchestrator reliability, async/mock compatibility, AI enhancement integration, fallback behavior, domain model backward compatibility, memory service legacy APIs, and security/rules fixes across the staged backend files.

+++ backend/src/services/chat/request_controls.py
@@ -135,10 +135,10 @@ class ChatRateLimiter:
                 $3,
                 $4,
                 1,
-                $5,
-                $6,
+                $5::integer,
+                $6::integer,
                 NOW(),
-                NOW() + ($6 * INTERVAL '1 second'),
+                NOW() + ($6::integer * INTERVAL '1 second'),
                 NOW()
             )
             ON CONFLICT (bucket_key)
